Process linter corrections

+++ vehicle/raw_vehicle_cmd_converter/src/interpolate.cpp
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <vector>
+
 #include "raw_vehicle_cmd_converter/interpolate.hpp"
 
 /*
@@ -23,7 +25,7 @@ bool LinearInterpolate::interpolate(
   const double & return_index, double & return_value)
 {
   auto isIncrease = [](const std::vector<double> & x) {
-      for (int i = 0; i < (int)x.size() - 1; ++i) {
+      for (size_t i = 0; i < x.size() - 1; ++i) {
         if (x[i] > x[i + 1]) {return false;}
       }
       return true;
@@ -65,7 +67,7 @@ bool LinearInterpolate::interpolate(
   }
 
   // calculate linear interpolation
-  int i = 0;
+  size_t i = 0;
   if (base_index[i] == return_index) {
     return_value = base_value[i];
     return true;
@@ -73,7 +75,7 @@ bool LinearInterpolate::interpolate(
   while (base_index[i] < return_index) {
     ++i;
   }
-  if (i <= 0 || (int)base_index.size() - 1 < i) {
+  if (i <= 0 || base_index.size() - 1 < i) {
     std::cerr << "? something wrong. skip this return_index." << std::endl;
     return false;
   }
